requirements.txt
rouge
mauve-text
bert-score
pot
sentence-transformers
geomloss
optimum 
sentencepiece
datasets
astropy
transformers
torch
nltk
fastdtw
torchmetrics
pyemd